refactor: purgomalumeClient class 위치를 domain package로 이동
