+++ src/main/java/kitchenpos/products/tobe/domain/ProductName.java
@@ -1,7 +1,5 @@
 package kitchenpos.products.tobe.domain;
 
-import kitchenpos.products.infra.purgomalum.PurgomalumClient;
-
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 import java.util.Objects;
