+++ src/main/java/kitchenpos/products/tobe/domain/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.products.infra.purgomalum;
+package kitchenpos.products.tobe.domain;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
