+++ src/test/java/kitchenpos/menus/application/MenuServiceTest.java
@@ -8,7 +8,7 @@ import kitchenpos.products.application.FakePurgomalumClient;
 import kitchenpos.products.application.InMemoryProductRepository;
 import kitchenpos.products.tobe.domain.Product;
 import kitchenpos.products.tobe.domain.ProductRepository;
-import kitchenpos.products.infra.purgomalum.PurgomalumClient;
+import kitchenpos.products.tobe.domain.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
