+++ src/test/java/kitchenpos/products/application/FakePurgomalumClient.java
@@ -1,6 +1,6 @@
 package kitchenpos.products.application;
 
-import kitchenpos.products.infra.purgomalum.PurgomalumClient;
+import kitchenpos.products.tobe.domain.PurgomalumClient;
 
 import java.util.Arrays;
 import java.util.List;
