+++ src/test/java/kitchenpos/products/application/ProductServiceTest.java
@@ -3,7 +3,7 @@ package kitchenpos.products.application;
 import kitchenpos.products.application.dto.ProductRequest;
 import kitchenpos.products.application.dto.ProductResponse;
 import kitchenpos.products.tobe.domain.ProductRepository;
-import kitchenpos.products.infra.purgomalum.PurgomalumClient;
+import kitchenpos.products.tobe.domain.PurgomalumClient;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
